data-pipeline/requirements-py313.txt
# data-pipeline/requirements-py313.txt
# Python dependencies for AirTable to PostgreSQL synchronization
# Version constraints relaxed for Python 3.13 compatibility

# PostgreSQL database adapter - letting pip choose compatible version
psycopg2-binary

# HTTP requests library - generally very stable across Python versions
requests>=2.31.0

# Let pip find pandas version that works with Python 3.13
pandas

# These packages typically work across Python versions
python-dateutil>=2.8.0
python-dotenv>=1.0.0
tenacity>=8.2.0
schedule>=1.2.0
tqdm>=4.66.0
jsonschema>=4.20.0
redis>=5.0.0
prometheus-client>=0.19.0

# NumPy is a pandas dependency but specifying it explicitly
# helps pip resolve dependencies more efficiently
numpy
